Use ipaddress module instead of ipaddr
ipaddress is part of the Python 3 standard lib, and is a more
recent fork of ipaddr. A Python 2 backport is available in pypi
if anyone needs to run this on Python 2 for some reason. ipaddr
upstream is not compatible with Python 3, and hasn't merged a
pull request for Python 3 compatibility that's been open for
several months.